StudyVerse-main/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-Mail==0.9.1
Werkzeug==3.0.1
requests==2.31.0
# google-generativeai==0.3.2
PyPDF2==3.0.1
openai==1.3.7
python-dotenv==1.0.0
authlib==1.3.0
pytz==2024.1
psycopg2-binary==2.9.11
gunicorn==21.2.0
eventlet>=0.35.2
python-socketio==5.12.0
dnspython==2.4.2
simple-websocket==1.0.0
whitenoise==6.6.0